Hice que muestre todos los avisos y que si estan resueltos se pueda eliminar

+++ Aulas/src/main/java/com/TrabajoFinal/Aulas/service/AvisoService.java
@@ -68,6 +68,11 @@ public class AvisoService {
 
     public void borrarAviso(Integer id){
         Aviso borrado = avisoXid(id);
+
+        if (!borrado.getEstado().equals(Estado.RESUELTO)) {
+            throw new RuntimeException("Solo se pueden eliminar avisos resueltos.");
+        }
+
         avisoRepository.delete(borrado);
     }
